+++ yaml/_loader.ts
@@ -39,15 +39,7 @@ import { YamlError } from "./_error.ts";
 import { Mark } from "./_mark.ts";
 import { DEFAULT_SCHEMA, type Schema, type TypeMap } from "./_schema.ts";
 import type { Type } from "./_type.ts";
-import {
-  type Any,
-  type ArrayObject,
-  codepointToChar,
-  decimalCharCodeToNumber,
-  getObjectTypeString,
-  hexCharCodeToNumber,
-  isObject,
-} from "./_utils.ts";
+import { type Any, type ArrayObject, isObject } from "./_utils.ts";
 
 const CONTEXT_FLOW_IN = 1;
 const CONTEXT_FLOW_OUT = 2;
@@ -78,6 +70,77 @@ interface LoaderStateOptions {
 
 // deno-lint-ignore no-explicit-any
 type ResultType = any[] | Record<string, any> | string;
+
+const ESCAPED_HEX_LENGTHS = new Map<number, number>([
+  [0x78, 2], // x
+  [0x75, 4], // u
+  [0x55, 8], // U
+]);
+
+const SIMPLE_ESCAPE_SEQUENCES = new Map<number, string>([
+  [0x30, "\x00"], // 0
+  [0x61, "\x07"], // a
+  [0x62, "\x08"], // b
+  [0x74, "\x09"], // t
+  [0x09, "\x09"], // Tab
+  [0x6e, "\x0A"], // n
+  [0x76, "\x0B"], // v
+  [0x66, "\x0C"], // f
+  [0x72, "\x0D"], // r
+  [0x65, "\x1B"], // e
+  [0x20, " "], // Space
+  [0x22, '"'], // "
+  [0x2f, "/"], // /
+  [0x5c, "\\"], // \
+  [0x4e, "\x85"], // N
+  [0x5f, "\xA0"], // _
+  [0x4c, "\u2028"], // L
+  [0x50, "\u2029"], // P
+]);
+
+function getObjectTypeString(object: unknown) {
+  return Object.prototype.toString.call(object);
+}
+
+/**
+ * Converts a hexadecimal character code to its decimal value.
+ */
+function hexCharCodeToNumber(charCode: number) {
+  // Check if the character code is in the range for '0' to '9'
+  if (0x30 <= charCode && charCode <= 0x39) return charCode - 0x30; // Convert '0'-'9' to 0-9
+
+  // Normalize the character code to lowercase if it's a letter
+  const lc = charCode | 0x20;
+
+  // Check if the character code is in the range for 'a' to 'f'
+  if (0x61 <= lc && lc <= 0x66) return lc - 0x61 + 10; // Convert 'a'-'f' to 10-15
+
+  return -1;
+}
+
+/**
+ * Converts a decimal character code to its decimal value.
+ */
+function decimalCharCodeToNumber(charCode: number): number {
+  // Check if the character code is in the range for '0' to '9'
+  if (0x30 <= charCode && charCode <= 0x39) return charCode - 0x30; // Convert '0'-'9' to 0-9
+  return -1;
+}
+
+/**
+ * Converts a Unicode code point to a string.
+ */
+function codepointToChar(codepoint: number): string {
+  // Check if the code point is within the Basic Multilingual Plane (BMP)
+  if (codepoint <= 0xffff) return String.fromCharCode(codepoint); // Convert BMP code point to character
+
+  // Encode UTF-16 surrogate pair for code points beyond BMP
+  // Reference: https://en.wikipedia.org/wiki/UTF-16#Code_points_U.2B010000_to_U.2B10FFFF
+  return String.fromCharCode(
+    ((codepoint - 0x010000) >> 10) + 0xd800, // High surrogate
+    ((codepoint - 0x010000) & 0x03ff) + 0xdc00, // Low surrogate
+  );
+}
 
 class LoaderState {
   schema: Schema;
@@ -154,33 +217,6 @@ class LoaderState {
     this.onWarning?.(error);
   }
 }
-
-const ESCAPED_HEX_LENGTHS = new Map<number, number>([
-  [0x78, 2], // x
-  [0x75, 4], // u
-  [0x55, 8], // U
-]);
-
-const SIMPLE_ESCAPE_SEQUENCES = new Map<number, string>([
-  [0x30, "\x00"], // 0
-  [0x61, "\x07"], // a
-  [0x62, "\x08"], // b
-  [0x74, "\x09"], // t
-  [0x09, "\x09"], // Tab
-  [0x6e, "\x0A"], // n
-  [0x76, "\x0B"], // v
-  [0x66, "\x0C"], // f
-  [0x72, "\x0D"], // r
-  [0x65, "\x1B"], // e
-  [0x20, " "], // Space
-  [0x22, '"'], // "
-  [0x2f, "/"], // /
-  [0x5c, "\\"], // \
-  [0x4e, "\x85"], // N
-  [0x5f, "\xA0"], // _
-  [0x4c, "\u2028"], // L
-  [0x50, "\u2029"], // P
-]);
 
 function yamlDirectiveHandler(state: LoaderState, ...args: string[]) {
   if (state.version !== null) {
